backend/file-upload/requirements.txt
psycopg2-binary==2.9.9
pyjwt==2.8.0
